mcp: address review — redact panic logs, relocate write-timeout test (#362)
CodeRabbit:
- Log recovered panic TYPE (%T) + stack instead of %v of the value in both
  the response-write recover and the #357 handler recover — a panic value
  can carry arbitrary request/tool payloads; the stack already locates the
  fault.
- Move the WriteTimeout assertion out of a NEW internal *_test.go file (per
  AGENTS.md: no new test files under internal/) into the existing
  server_internal_test.go. It must stay in-package because newMCPHTTPServer
  is unexported and unreachable from tests/.
